refactor: extract i18n from languageHandler to not import matrix-js-sdk, settings...
